Unify conversation archive, delete and block use cases
