+++ src/com/android/messaging/data/blockedparticipants/repository/BlockedParticipantsRepository.kt
@@ -10,9 +10,13 @@ import com.android.messaging.datamodel.DatabaseHelper
 import com.android.messaging.datamodel.DatabaseHelper.ConversationColumns
 import com.android.messaging.datamodel.DatabaseHelper.ParticipantColumns
 import com.android.messaging.datamodel.MessagingContentProvider
+import com.android.messaging.datamodel.action.UpdateDestinationBlockedAction.UpdateDestinationBlockedActionListener
+import com.android.messaging.datamodel.action.UpdateDestinationBlockedAction.updateDestinationBlocked
 import com.android.messaging.datamodel.data.ParticipantData
+import com.android.messaging.di.core.MainDispatcher
 import com.android.messaging.di.core.MessagingDbDispatcher
 import javax.inject.Inject
+import kotlin.coroutines.resume
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
@@ -22,14 +26,23 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlinx.coroutines.withContext
 
 internal interface BlockedParticipantsRepository {
     fun observeBlockedParticipants(): Flow<ImmutableList<BlockedDirectChat>>
+
+    suspend fun setDestinationBlocked(
+        destination: String,
+        conversationId: String?,
+        isBlocked: Boolean,
+    ): Boolean
 }
 
 internal class BlockedParticipantsRepositoryImpl @Inject constructor(
     private val contentResolver: ContentResolver,
     @param:MessagingDbDispatcher private val messagingDbDispatcher: CoroutineDispatcher,
+    @param:MainDispatcher private val mainDispatcher: CoroutineDispatcher,
 ) : BlockedParticipantsRepository {
 
     private val dataModel = DataModel.get()
@@ -43,6 +56,35 @@ internal class BlockedParticipantsRepositoryImpl @Inject constructor(
         return observeUris(uris)
             .map { queryBlockedDirectChats() }
             .flowOn(messagingDbDispatcher)
+    }
+
+    override suspend fun setDestinationBlocked(
+        destination: String,
+        conversationId: String?,
+        isBlocked: Boolean,
+    ): Boolean {
+        val resolvedDestination = destination.takeIf(String::isNotBlank) ?: return false
+
+        return withContext(mainDispatcher) {
+            suspendCancellableCoroutine { continuation ->
+                val listener = UpdateDestinationBlockedActionListener { _, success, _, _ ->
+                    if (continuation.isActive) {
+                        continuation.resume(success)
+                    }
+                }
+
+                val actionMonitor = updateDestinationBlocked(
+                    resolvedDestination,
+                    isBlocked,
+                    conversationId?.takeIf(String::isNotBlank),
+                    listener,
+                )
+
+                continuation.invokeOnCancellation {
+                    actionMonitor?.unregister()
+                }
+            }
+        }
     }
 
     // Returns only blocked participants that have an existing 1 on 1 conversation.
