+++ src/com/android/messaging/di/conversationlist/ConversationListBindsModule.kt
@@ -4,12 +4,6 @@ import com.android.messaging.data.conversationlist.repository.ConversationListRe
 import com.android.messaging.data.conversationlist.repository.ConversationListRepositoryImpl
 import com.android.messaging.data.conversationlist.store.ConversationListStatusStore
 import com.android.messaging.data.conversationlist.store.ConversationListStatusStoreImpl
-import com.android.messaging.domain.conversationlist.usecase.DeleteConversations
-import com.android.messaging.domain.conversationlist.usecase.DeleteConversationsImpl
-import com.android.messaging.domain.conversationlist.usecase.SetConversationArchived
-import com.android.messaging.domain.conversationlist.usecase.SetConversationArchivedImpl
-import com.android.messaging.domain.conversationlist.usecase.SetConversationBlocked
-import com.android.messaging.domain.conversationlist.usecase.SetConversationBlockedImpl
 import com.android.messaging.ui.conversationlist.redesign.mapper.ConversationListUiStateMapper
 import com.android.messaging.ui.conversationlist.redesign.mapper.ConversationListUiStateMapperImpl
 import dagger.Binds
@@ -39,22 +33,4 @@ internal abstract class ConversationListBindsModule {
     abstract fun bindConversationListUiStateMapper(
         impl: ConversationListUiStateMapperImpl,
     ): ConversationListUiStateMapper
-
-    @Binds
-    @Reusable
-    abstract fun bindDeleteConversations(
-        impl: DeleteConversationsImpl,
-    ): DeleteConversations
-
-    @Binds
-    @Reusable
-    abstract fun bindSetConversationArchived(
-        impl: SetConversationArchivedImpl,
-    ): SetConversationArchived
-
-    @Binds
-    @Reusable
-    abstract fun bindSetConversationBlocked(
-        impl: SetConversationBlockedImpl,
-    ): SetConversationBlocked
 }
