--- src/com/android/messaging/domain/conversationlist/model/ConversationListActionTarget.kt
@@ -1,6 +1,0 @@
-package com.android.messaging.domain.conversationlist.model
-
-internal data class ConversationListActionTarget(
-    val conversationId: String,
-    val cutoffTimestampMillis: Long,
-)
